Extracted ExerciseInputFields
